Refactor Script constructors to avoid fromComponents()
Script.fromComponents() is good for keeping you from shooting yourself
in the foot, but it does add overhead with all the stream and future
forwarding. Since we know what we're doing (in particular, we know to
manually check the capture state before doing side effects), we can
avoid that overhead.